+++ server/src/main/java/org/allaymc/server/entity/effect/EffectResistanceType.java
@@ -2,6 +2,7 @@ package org.allaymc.server.entity.effect;
 
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.damage.DamageContainer;
+import org.allaymc.api.entity.damage.DamageType;
 import org.allaymc.api.entity.effect.AbstractEffectType;
 import org.allaymc.api.entity.effect.EffectInstance;
 import org.allaymc.api.entity.effect.EffectTypes;
@@ -20,9 +21,9 @@ public class EffectResistanceType extends AbstractEffectType {
     @Override
     public void onEntityDamage(Entity entity, EffectInstance effectInstance, DamageContainer damage) {
         if (
-                damage.getDamageType() == DamageContainer.DamageType.VOID ||
-                damage.getDamageType() == DamageContainer.DamageType.STARVE ||
-                damage.getDamageType() == DamageContainer.DamageType.COMMAND
+                damage.getDamageType() == DamageType.VOID ||
+                damage.getDamageType() == DamageType.STARVE ||
+                damage.getDamageType() == DamageType.COMMAND
         ) return;
 
         var level = entity.getEffectLevel(EffectTypes.RESISTANCE);